refactor: betAmount 디렉토리 이동
