+++ src/main/java/controller/BlackjackController.java
@@ -7,10 +7,10 @@ import domain.blackjackgame.BlackjackGame;
 import domain.blackjackgame.GameResult;
 import domain.card.deck.CardDeck;
 import domain.card.deck.CardDeckFactory;
-import domain.participant.BetAmount;
 import domain.participant.Dealer;
 import domain.participant.Participants;
 import domain.participant.Player;
+import domain.participant.betting.BetAmount;
 import domain.participant.name.Name;
 import domain.participant.name.Names;
 import java.util.Collections;
