+++ src/test/java/domain/participant/betting/BetAmountTest.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.participant.betting;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
